@@ -1,10 +1,9 @@
-﻿namespace AuthenticationProject.Models
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace AuthenticationProject.Models;
+
+public class User : IdentityUser
 {
-    public class User
-    {
-        public int Id { get; set; }
-        public string Username { get; set; }
-        public string Password { get; set; }
-        public string Role { get; set; }
-    }
+    public string FirsName { get; set; }
+    public string LastName { get; set; }
 }
