@@ -1,12 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace AuthenticationProject;
 
-namespace AuthenticationProject
+public class Settings
 {
-    public class Settings
-    {
-        public static string Secret = "ec0e2603172c73a8b644bb9456c1ff6e";
-    }
+    public static string Secret = "ec0e2603172c73a8b644bb9456c1ff6e";
 }
